requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
numpy==1.26.3 